perf: store emebdding inside memmap
