+++ jina/types/arrays/document.py
@@ -17,8 +17,6 @@ from typing import (
     Sequence,
 )
 
-import numpy as np
-
 from .neural_ops import DocumentArrayNeuralOpsMixin
 from .search_ops import DocumentArraySearchOpsMixin
 from .traversable import TraversableSequence
@@ -46,63 +44,6 @@ DocumentArraySourceType = TypeVar(
     Sequence[jina_pb2.DocumentProto],
     Document,
 )
-
-
-class DocumentArrayMemmapGetAttrMixin:
-    """A mixin that provides attributes getter in bulk """
-
-    @abstractmethod
-    def __iter__(self):
-        ...
-
-    def get_attributes(self, *fields: str) -> Union[List, List[List]]:
-        """Return all nonempty values of the fields from all docs this array contains
-
-        :param fields: Variable length argument with the name of the fields to extract
-        :return: Returns a list of the values for these fields.
-            When `fields` has multiple values, then it returns a list of list.
-        """
-        fields = list(fields)
-        if 'embedding' in fields and not self._is_updated:
-            embeddings = self.embeddings_memmap.tolist()
-            index = fields.index('embedding')
-            fields.remove('embedding')
-        elif 'embedding' in fields and self._is_updated:
-            embeddings = [doc.get_attributes('embedding') for doc in self]
-            self.embeddings_memmap = np.asarray(embeddings)
-            index = fields.index('embedding')
-            fields.remove('embedding')
-        if fields:
-            contents = [doc.get_attributes(*fields) for doc in self]
-            if len(fields) > 1:
-                contents = list(map(list, zip(*contents)))
-            if index:
-                contents.insert(index, embeddings)
-
-        return contents
-
-    def get_attributes_with_docs(
-        self,
-        *fields: str,
-    ) -> Tuple[Union[List, List[List]], 'DocumentArray']:
-        """Return all nonempty values of the fields together with their nonempty docs
-
-        :param fields: Variable length argument with the name of the fields to extract
-        :return: Returns a tuple. The first element is  a list of the values for these fields.
-            When `fields` has multiple values, then it returns a list of list. The second element is the non-empty docs.
-        """
-
-        contents = []
-        docs_pts = []
-
-        for doc in self:
-            contents.append(doc.get_attributes(*fields))
-            docs_pts.append(doc)
-
-        if len(fields) > 1:
-            contents = list(map(list, zip(*contents)))
-
-        return contents, DocumentArray(docs_pts)
 
 
 class DocumentArrayGetAttrMixin:
